feat(type): 제품 관련 타입 분리
